+++ paddle_billing/Entities/Subscriptions/__init__.py
@@ -2,7 +2,6 @@ from paddle_billing.Entities.Subscriptions.SubscriptionAdjustmentItem           
 from paddle_billing.Entities.Subscriptions.SubscriptionAdjustmentPreview                import SubscriptionAdjustmentPreview
 from paddle_billing.Entities.Subscriptions.SubscriptionCharge                           import SubscriptionCharge
 from paddle_billing.Entities.Subscriptions.SubscriptionCredit                           import SubscriptionCredit
-from paddle_billing.Entities.Subscriptions.SubscriptionDetails                          import SubscriptionDetails
 from paddle_billing.Entities.Subscriptions.SubscriptionDiscount                         import SubscriptionDiscount
 from paddle_billing.Entities.Subscriptions.SubscriptionEffectiveFrom                    import SubscriptionEffectiveFrom
 from paddle_billing.Entities.Subscriptions.SubscriptionItem                             import SubscriptionItem
@@ -16,7 +15,6 @@ from paddle_billing.Entities.Subscriptions.SubscriptionNonCatalogPriceWithProduc
 from paddle_billing.Entities.Subscriptions.SubscriptionNonCatalogProduct                import SubscriptionNonCatalogProduct
 from paddle_billing.Entities.Subscriptions.SubscriptionOnPaymentFailure                 import SubscriptionOnPaymentFailure
 from paddle_billing.Entities.Subscriptions.SubscriptionPreviewSubscriptionUpdateSummary import SubscriptionPreviewSubscriptionUpdateSummary
-from paddle_billing.Entities.Subscriptions.SubscriptionProration                        import SubscriptionProration
 from paddle_billing.Entities.Subscriptions.SubscriptionProrationBillingMode             import SubscriptionProrationBillingMode
 from paddle_billing.Entities.Subscriptions.SubscriptionResult                           import SubscriptionResult
 from paddle_billing.Entities.Subscriptions.SubscriptionResultAction                     import SubscriptionResultAction
@@ -24,7 +22,4 @@ from paddle_billing.Entities.Subscriptions.SubscriptionResumeEffectiveFrom      
 from paddle_billing.Entities.Subscriptions.SubscriptionScheduledChange                  import SubscriptionScheduledChange
 from paddle_billing.Entities.Subscriptions.SubscriptionScheduledChangeAction            import SubscriptionScheduledChangeAction
 from paddle_billing.Entities.Subscriptions.SubscriptionStatus                           import SubscriptionStatus
-from paddle_billing.Entities.Subscriptions.SubscriptionTimePeriod                       import SubscriptionTimePeriod
-from paddle_billing.Entities.Subscriptions.SubscriptionTransactionItem                  import SubscriptionTransactionItem
-from paddle_billing.Entities.Subscriptions.SubscriptionTransactionLineItem              import SubscriptionTransactionLineItem
 from paddle_billing.Entities.Subscriptions.SubscriptionUpdateItem                       import SubscriptionUpdateItem