+++ paddle_billing/Entities/Transaction.py
@@ -16,6 +16,7 @@ from paddle_billing.Entities.Shared import (
     CurrencyCode,
     CustomData,
     PaymentMethodType,
+    TimePeriod,
     TransactionOrigin,
     TransactionPaymentAttempt,
     TransactionStatus,
@@ -24,7 +25,6 @@ from paddle_billing.Entities.Shared import (
 from paddle_billing.Entities.Transactions.TransactionAdjustmentsTotals import TransactionAdjustmentsTotals
 from paddle_billing.Entities.Transactions.TransactionDetails           import TransactionDetails
 from paddle_billing.Entities.Transactions.TransactionItem              import TransactionItem
-from paddle_billing.Entities.Transactions.TransactionTimePeriod        import TransactionTimePeriod
 
 
 @dataclass
@@ -41,9 +41,9 @@ class Transaction(Entity):
     invoice_id:                str | None
     invoice_number:            str | None
     collection_mode:           CollectionMode
-    discount_id:               str                   | None
-    billing_details:           BillingDetails        | None
-    billing_period:            TransactionTimePeriod | None
+    discount_id:               str            | None
+    billing_details:           BillingDetails | None
+    billing_period:            TimePeriod     | None
     items:                     list[TransactionItem]
     details:                   TransactionDetails
     payments:                  list[TransactionPaymentAttempt]
@@ -84,7 +84,7 @@ class Transaction(Entity):
             address         = Address.from_dict(data['address'])                      if data.get('address')         else None,
             billed_at       = datetime.fromisoformat(data['billed_at'])               if data.get('billed_at')       else None,
             billing_details = BillingDetails.from_dict(data['billing_details'])       if data.get('billing_details') else None,
-            billing_period  = TransactionTimePeriod.from_dict(data['billing_period']) if data.get('billing_period')  else None,
+            billing_period  = TimePeriod.from_dict(data['billing_period'])            if data.get('billing_period')  else None,
             business        = Business.from_dict(data['business'])                    if data.get('business')        else None,
             custom_data     = CustomData(data['custom_data'])                         if data.get('custom_data')     else None,
             checkout        = Checkout.from_dict(data['checkout'])                    if data.get('checkout')        else None,