+++ paddle_billing/Notifications/Entities/Subscriptions/SubscriptionItem.py
@@ -2,10 +2,11 @@ from __future__  import annotations
 from dataclasses import dataclass
 from datetime    import datetime
 
+from paddle_billing.Notifications.Entities.Shared.TimePeriod import TimePeriod
+
 from paddle_billing.Notifications.Entities.Product import Product
 from paddle_billing.Notifications.Entities.Subscriptions.SubscriptionItemStatus import SubscriptionItemStatus
 from paddle_billing.Notifications.Entities.Subscriptions.SubscriptionPrice      import SubscriptionPrice
-from paddle_billing.Notifications.Entities.Subscriptions.SubscriptionTimePeriod import SubscriptionTimePeriod
 
 
 @dataclass
@@ -17,7 +18,7 @@ class SubscriptionItem:
     updated_at:           datetime
     previously_billed_at: datetime | None
     next_billed_at:       datetime | None
-    trial_dates:          SubscriptionTimePeriod | None
+    trial_dates:          TimePeriod | None
     price:                SubscriptionPrice
     product:              Product | None
 
@@ -33,6 +34,6 @@ class SubscriptionItem:
             price                = SubscriptionPrice.from_dict(data['price']),
             previously_billed_at = datetime.fromisoformat(data['previously_billed_at'])  if data.get('previously_billed_at') else None,
             next_billed_at       = datetime.fromisoformat(data['next_billed_at'])        if data.get('next_billed_at')       else None,
-            trial_dates          = SubscriptionTimePeriod.from_dict(data['trial_dates']) if data.get('trial_dates')          else None,
+            trial_dates          = TimePeriod.from_dict(data['trial_dates']) if data.get('trial_dates')          else None,
             product              = Product.from_dict(data['product'])                    if data.get('product')          else None,
         )