+++ paddle_billing/Notifications/Entities/Transactions/__init__.py
@@ -1,8 +1,6 @@
 from paddle_billing.Notifications.Entities.Transactions.TransactionAdjustmentsTotals import TransactionAdjustmentsTotals
 from paddle_billing.Notifications.Entities.Transactions.TransactionBreakdown         import TransactionBreakdown
 from paddle_billing.Notifications.Entities.Transactions.TransactionCardType          import TransactionCardType
-from paddle_billing.Notifications.Entities.Transactions.TransactionProration         import TransactionProration
-from paddle_billing.Notifications.Entities.Transactions.TransactionTimePeriod        import TransactionTimePeriod
 
 
 # These cause circular imports