+++ paddle_billing/Resources/Prices/Operations/UpdatePrice.py
@@ -1,7 +1,7 @@
 from dataclasses import asdict, dataclass
 
 from paddle_billing.Undefined       import Undefined
-from paddle_billing.Entities.Shared import CatalogType, CustomData, Money, PriceQuantity, Status, TaxMode, TimePeriod, UnitPriceOverride
+from paddle_billing.Entities.Shared import CatalogType, CustomData, Duration, Money, PriceQuantity, Status, TaxMode, UnitPriceOverride
 
 
 @dataclass
@@ -9,8 +9,8 @@ class UpdatePrice:
     description:          str                            | Undefined = Undefined()
     name:                 str                     | None | Undefined = Undefined()
     type:                 CatalogType                    | Undefined = Undefined()
-    billing_cycle:        TimePeriod              | None | Undefined = Undefined()
-    trial_period:         TimePeriod              | None | Undefined = Undefined()
+    billing_cycle:        Duration                | None | Undefined = Undefined()
+    trial_period:         Duration                | None | Undefined = Undefined()
     tax_mode:             TaxMode                        | Undefined = Undefined()
     unit_price:           Money                          | Undefined = Undefined()
     unit_price_overrides: list[UnitPriceOverride]        | Undefined = Undefined()